feat: put clothes and accessories to js filesconst accessories = [
