+++ src/js/accessories.js
@@ -1,0 +1,46 @@
+const accessories = [
+    {
+        id:    7,
+        title: 'Кулон "Green"',
+        price: 250,
+        isNew: true,
+        img:   'src/assets/img/accessory1.jpg',
+    },
+    {
+        id:    8,
+        title: 'Браслет "Gold"',
+        price: 270,
+        isNew: false,
+        img:   'src/assets/img/accessory2.jpg',
+    },
+    {
+        id:    9,
+        title: 'Часы "Omega"',
+        price: 200,
+        isNew: false,
+        img:   'src/assets/img/accessory3.jpg',
+    },
+    {
+        id:    10,
+        title: 'Селфи палка "Blue"',
+        price: 180,
+        isNew: true,
+        img:   'src/assets/img/accessory4.jpg',
+    },
+    {
+        id:    11,
+        title: 'Бутылка "Versace"',
+        price: 1000,
+        isNew: true,
+        img:   'src/assets/img/accessory5.jpg',
+    },
+    {
+        id:    11,
+        title: 'Очки "Black"',
+        price: 170,
+        isNew: false,
+        img:   'src/assets/img/accessory6.jpg',
+    },
+];
+
+export default accessories;
