+++ src/js/clothes.js
@@ -1,0 +1,53 @@
+const clothes = [
+    {
+        id:    0,
+        title: 'Футболка "Эволюционируй или Сдохни"',
+        price: 220,
+        isNew: true,
+        img:   'src/assets/img/goods-item.png',
+    },
+    {
+        id:    1,
+        title: 'Футболка "There is always second chance"',
+        price: 310,
+        isNew: true,
+        img:   'src/assets/img/cloth1.jpg',
+    },
+    {
+        id:    2,
+        title: 'Футболка "Brown"',
+        price: 280,
+        isNew: false,
+        img:   'src/assets/img/cloth2.jpg',
+    },
+    {
+        id:    3,
+        title: 'Футболка "Pringles"',
+        price: 480,
+        isNew: false,
+        img:   'src/assets/img/cloth3.jpg',
+    },
+    {
+        id:    4,
+        title: 'Футболка "Мои кореша"',
+        price: 280,
+        isNew: true,
+        img:   'src/assets/img/cloth4.jpg',
+    },
+    {
+        id:    5,
+        title: 'Футболка "KEU"',
+        price: 230,
+        isNew: false,
+        img:   'src/assets/img/cloth5.jpg',
+    },
+    {
+        id:    6,
+        title: 'Футболка "Bella Ciao"',
+        price: 150,
+        isNew: false,
+        img:   'src/assets/img/cloth6.jpg',
+    },
+];
+
+export default clothes;
