+++ src/js/main.js
@@ -1,4 +1,6 @@
 import '../scss/style.scss';
+import clothes from './clothes';
+import accessories from './accessories';
 
 const categories = document.querySelectorAll('.js__category-button');
 const goodsItems = document.querySelectorAll('.goods__item');
@@ -201,103 +203,6 @@ class GoodsItem {
         this.parent.append(element);
     }
 }
-
-const clothes = [
-    {
-        id:    0,
-        title: 'Футболка "Эволюционируй или Сдохни"',
-        price: 220,
-        isNew: true,
-        img:   'src/assets/img/goods-item.png',
-    },
-    {
-        id:    1,
-        title: 'Футболка "There is always second chance"',
-        price: 310,
-        isNew: true,
-        img:   'src/assets/img/cloth1.jpg',
-    },
-    {
-        id:    2,
-        title: 'Футболка "Brown"',
-        price: 280,
-        isNew: false,
-        img:   'src/assets/img/cloth2.jpg',
-    },
-    {
-        id:    3,
-        title: 'Футболка "Pringles"',
-        price: 480,
-        isNew: false,
-        img:   'src/assets/img/cloth3.jpg',
-    },
-    {
-        id:    4,
-        title: 'Футболка "Мои кореша"',
-        price: 280,
-        isNew: true,
-        img:   'src/assets/img/cloth4.jpg',
-    },
-    {
-        id:    5,
-        title: 'Футболка "KEU"',
-        price: 230,
-        isNew: false,
-        img:   'src/assets/img/cloth5.jpg',
-    },
-    {
-        id:    6,
-        title: 'Футболка "Bella Ciao"',
-        price: 150,
-        isNew: false,
-        img:   'src/assets/img/cloth6.jpg',
-    },
-];
-
-const accessories = [
-    {
-        id:    7,
-        title: 'Кулон "Green"',
-        price: 250,
-        isNew: true,
-        img:   'src/assets/img/accessory1.jpg',
-    },
-    {
-        id:    8,
-        title: 'Браслет "Gold"',
-        price: 270,
-        isNew: false,
-        img:   'src/assets/img/accessory2.jpg',
-    },
-    {
-        id:    9,
-        title: 'Часы "Omega"',
-        price: 200,
-        isNew: false,
-        img:   'src/assets/img/accessory3.jpg',
-    },
-    {
-        id:    10,
-        title: 'Селфи палка "Blue"',
-        price: 180,
-        isNew: true,
-        img:   'src/assets/img/accessory4.jpg',
-    },
-    {
-        id:    11,
-        title: 'Бутылка "Versace"',
-        price: 1000,
-        isNew: true,
-        img:   'src/assets/img/accessory5.jpg',
-    },
-    {
-        id:    11,
-        title: 'Очки "Black"',
-        price: 170,
-        isNew: false,
-        img:   'src/assets/img/accessory6.jpg',
-    },
-];
 
 const sortedCloth = clothes.sort((x, y) => y.isNew - x.isNew);
 const sortedAccessories = accessories.sort((x, y) => y.isNew - x.isNew);
